+++ app/src/main/java/com/disfluency/AppScaffold.kt
@@ -13,11 +13,12 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import com.disfluency.navigation.*
+import com.disfluency.navigation.bottomNavigation.BottomNavigation
+import com.disfluency.navigation.bottomNavigation.BottomNavigationItem
 
 @Composable
 fun AppScaffold(bottomNavigationItems: List<BottomNavigationItem>, content: @Composable (NavHostController)->Unit) {
