+++ app/src/main/java/com/disfluency/data/UserRepository.kt
@@ -1,7 +1,7 @@
-package com.disfluency.screens.utils
+package com.disfluency.data
 
-import com.disfluency.data.PatientRepository
 import com.disfluency.model.Phono
+import com.disfluency.model.User
 import kotlinx.coroutines.delay
 
 class UserNotFoundException: Exception()
@@ -11,8 +11,7 @@ object UserRepository {
         User("beto", "123", Phono())
     )
 
-    suspend fun findUser(username: String, password: String): User{
-        delay(300)
+    suspend fun findUser(username: String, password: String): User {
         return users.find { it.username == username && it.password == password } ?: throw UserNotFoundException()
     }
 }
