+++ app/src/main/java/com/disfluency/model/Patient.kt
@@ -1,6 +1,5 @@
 package com.disfluency.model
 
-import com.disfluency.screens.utils.Role
 import java.time.LocalDate
 import java.time.Period
 
@@ -14,7 +13,7 @@ data class Patient(
     val profilePic: Int,
     val weeklyTurn: String, //TODO: ver cual seria el tipo de dato para esto
     val weeklyHour: String //TODO: ver cual seria el tipo de dato para esto
-    ): Role{
+    ): Role {
 
     fun initials(): String {
         return (name.first().toString() + lastName.first().toString()).uppercase()
