+++ app/src/main/java/com/disfluency/model/Phono.kt
@@ -1,7 +1,5 @@
 package com.disfluency.model
 
-import com.disfluency.screens.utils.Role
-
 class Phono: Role {
 
 }
