+++ app/src/main/java/com/disfluency/navigation/bottomNavigation/BottomNavigation.kt
@@ -1,0 +1,37 @@
+package com.disfluency.navigation.bottomNavigation
+
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.navigation.NavController
+import androidx.navigation.compose.currentBackStackEntryAsState
+
+@Composable
+fun BottomNavigation(navController: NavController, items: List<BottomNavigationItem>) {
+
+    NavigationBar {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentRoute = navBackStackEntry?.destination?.route
+
+        items.forEach { item ->
+
+            NavigationBarItem(
+                selected = currentRoute == item.screenRoute.route,
+                icon = { Icon(item.icon, contentDescription = item.screenRoute.title) },
+                label = { Text(text = item.screenRoute.title) },
+                onClick = {
+                    navController.navigate(item.screenRoute.route) {
+
+                        navController.graph.startDestinationRoute?.let { screen_route ->
+                            popUpTo(screen_route) {
+                                saveState = true
+                            }
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+}
