+++ app/src/main/java/com/disfluency/navigation/navigationGraphs/PatientNavigationGraph.kt
@@ -1,0 +1,24 @@
+package com.disfluency.navigation.navigationGraphs
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import com.disfluency.navigation.bottomNavigation.BottomNavigationItem
+import com.disfluency.navigation.Route
+import com.disfluency.screens.utils.EmptyScreen
+
+@Composable
+fun PatientNavigationGraph(navController: NavHostController, onLogout: ()->Unit) {
+    NavHost(navController, startDestination = Route.HomePatient.route) {
+        composable(BottomNavigationItem.HomePatient.screenRoute.route) {
+            EmptyScreen("Home Paciente", onLogout)
+        }
+        composable(BottomNavigationItem.Ejercicios.screenRoute.route) {
+            EmptyScreen("Ejercicios Paciente", onLogout)
+        }
+        composable(BottomNavigationItem.Cuestionarios.screenRoute.route) {
+            EmptyScreen("Cuestionarios Paciente", onLogout)
+        }
+    }
+}
