+++ app/src/main/java/com/disfluency/screens/phono/CuestionariosScreen.kt
@@ -1,4 +1,4 @@
-package com.disfluency.screens
+package com.disfluency.screens.phono
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,19 +11,16 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
-import com.disfluency.data.PatientRepository
 
 @Composable
-fun PatientExercisesScreen(id: Int) {
-    val patient = PatientRepository.getPatientById(id)
-
+fun PhonoCuestionariosScreen() {
     Column(
         modifier = Modifier
             .fillMaxSize()
             .wrapContentSize(Alignment.Center)
     ) {
         Text(
-            text = "Ejercicios del Paciente: ${patient.name}",
+            text = "Mis Cuestionarios",
             fontWeight = FontWeight.Bold,
             color = Color.Gray,
             modifier = Modifier.align(Alignment.CenterHorizontally),
