+++ app/src/main/java/com/disfluency/screens/phono/PatientSessionsScreen.kt
@@ -1,4 +1,4 @@
-package com.disfluency.screens
+package com.disfluency.screens.phono
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
@@ -11,16 +11,19 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.sp
+import com.disfluency.data.PatientRepository
 
 @Composable
-fun PhonoEjerciciosScreen() {
+fun PatientSessionsScreen(id: Int) {
+    val patient = PatientRepository.getPatientById(id)
+
     Column(
         modifier = Modifier
             .fillMaxSize()
             .wrapContentSize(Alignment.Center)
     ) {
         Text(
-            text = "Mis Ejercicios",
+            text = "Sesiones del Paciente: ${patient.name}",
             fontWeight = FontWeight.Bold,
             color = Color.Gray,
             modifier = Modifier.align(Alignment.CenterHorizontally),
