+++ app/src/main/java/com/disfluency/screens/utils/LoginService.kt
@@ -1,11 +1,15 @@
 package com.disfluency.screens.utils
 
+import com.disfluency.data.UserNotFoundException
+import com.disfluency.data.UserRepository
+import com.disfluency.model.User
 import kotlinx.coroutines.delay
 
 class LoginService {
     private var user: User? = null
 
     suspend fun login(username: String, password: String){
+        delay(50)
         val user = UserRepository.findUser(username, password)
         this.user = user
     }
@@ -15,7 +19,7 @@ class LoginService {
     }
 
     suspend fun logout(){
-        delay(300)
+        delay(50)
         this.user = null
     }
 }
